backend/menu-service/pyproject.toml
[project]
name = "menu-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Sivkim HENG",email = "hengsivkim14@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "cassandra-driver (>=3.29.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
menu-service = "uvicorn:main"
